Hard Shell (Alpha)
Hard Shell is a PowerShell script which by default changes password policies, creates a folder with client hardware information, etc. This is mainly used for system administrators, especially in deployment, but obviously I'm not here to tell you how to use it.

This is a very small project (obviously) and I don't expect a lot of people to use it.